pyproject.toml
[tool.poetry]
name = "gcp_mysql_backup_service"
version = "0.1.0"
description = "MySQL backup service"
authors = ["Viacheslav Bessonov <viacheslav.bessonov@hilbertteam.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-secret-manager = "^1.0.0"
google-cloud-storage = "^1.28.1"
google-cloud-pubsub = "^1.5.0"
"ruamel.yaml" = "^0.16.10"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
bumpversion = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
